Move extensions test into single file
